Изменение задач + рефакторинг (кстати, он для геев)

See merge request diploma_m_175/diploma!16

+++ server/src/main/java/com/server/task/model/Task.java
@@ -34,10 +34,10 @@ public class Task implements Serializable {
     private Long employeeId;
 
     @Column(name ="task_priority_id")
-    private int priorityId;
+    private Long priorityId;
 
     @Column(name = "task_status_id")
-    private int statusId;
+    private Long statusId;
 
     @ManyToMany(mappedBy = "tasks")
     private List<User> users = new ArrayList<>();
@@ -74,19 +74,19 @@ public class Task implements Serializable {
         this.parentId = null;
     }
 
-    public int getPriority() {
+    public Long getPriority() {
         return priorityId;
     }
 
-    public void setPriority(int priorityId) {
+    public void setPriority(Long priorityId) {
         this.priorityId = priorityId;
     }
 
-    public int getStatus() {
+    public Long getStatus() {
         return statusId;
     }
 
-    public void setStatus(int statusId) {
+    public void setStatus(Long statusId) {
         this.statusId = statusId;
     }
